+++ model/book.go
@@ -2,10 +2,10 @@ package model
 
 //Book holds the definition of a book
 type Book struct {
-	ID       int64   `json:"id" db:"book_id"`
+	ID       uint    `json:"id" db:"book_id"`
 	Name     string  `json:"name" db:"name"`
 	Text     string  `json:"text" db:"text"`
-	LawID    int64   `json:"lawID" db:"law_id"`
+	LawID    uint    `json:"lawID" db:"law_id"`
 	Titles   []Title `json:"titles"`
 	Reviewed bool    `json:"reviewed"`
 }
